% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{avgSilhoette}
\alias{avgSilhoette}
\title{compute the average value of silhouette coefficients}
\usage{
avgSilhoette(distMat, label)
}
\arguments{
\item{distMat}{a n x n matrix of pairwise distance between a set of n queries}

\item{label}{a vector of size n that contains the labels of n queries}
}
\value{
the average value of silhouette coefficients
}
\description{
compute the average value of silhouette coefficients
}
\author{
Duc Luong
}

