% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{BetaCV}
\alias{BetaCV}
\title{compute the BetaCV}
\usage{
BetaCV(distMat, label)
}
\arguments{
\item{distMat}{a n x n matrix of pairwise distance between a set of n queries}

\item{label}{a vector of size n that contains the labels of n queries}
}
\value{
a number that measure the DunnIndex of the pairwise distance matrix given the labels of queries
}
\description{
compute the BetaCV
}
\details{
The BetaCV measure is the ratio of the mean intracluster distance to the mean intercluster
distance. The smaller the BetaCV ratio, the better the clustering
}
\author{
Duc Luong
}
\references{
Zaki, Mohammed J., and Wagner Meira Jr. Data mining and analysis: fundamental concepts and algorithms. Cambridge University Press, 2014, Chapter 17.
}

