% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{computeCosinePairwiseDistance}
\alias{computeCosinePairwiseDistance}
\title{compute pairwise distance between queries from the set of feature vectors using cosine similarity}
\usage{
computeCosinePairwiseDistance(featureMat)
}
\arguments{
\item{featureMat}{a nxp matrix where n is the number of queries and p is the number of features}
}
\value{
a nxn matrix in which each entry contains a pairwise distance between queries
}
\description{
compute pairwise distance between queries from the set of feature vectors using cosine similarity
}
\author{
Duc Luong
}

