% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{computeDistMatTopic}
\alias{computeDistMatTopic}
\title{perform topic modeling using feature vector file and compare query similarity using Hellinger distance}
\usage{
computeDistMatTopic(labelFile, featureVecFile, datasetFile, numTopic,
  SEED = 2010)
}
\arguments{
\item{labelFile}{file that contains list labels}

\item{featureVecFile}{file that contains feature vectors in sparse representation (row, col, val). Each row is a feature vector for one corresponding query}

\item{datasetFile}{file that contains other information about queries (query itself, id, label)}

\item{numTopic}{number of topics}

\item{SEED}{the random seed for LDA using VEM algorithm}
}
\value{
the n x n matrix that contains distance between queries using Hellinger distance
}
\description{
perform topic modeling using feature vector file and compare query similarity using Hellinger distance
}
\author{
Duc Luong
}

