% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{minWeights}
\alias{minWeights}
\title{compute the minimum weight of all of the connections between two clusters}
\usage{
minWeights(distMat, label, i, j)
}
\arguments{
\item{distMat}{a n x n matrix of pairwise distance between a set of n queries}

\item{label}{a vector of size n that contains the labels of n queries}

\item{i, j}{cluster labels that we want to compute the minimum weight among all the edges between cluster i and cluster j}
}
\value{
the minimum weight among all connections between two clusters
}
\description{
compute the minimum weight of all of the connections between two clusters
}
\author{
Duc Luong
}

