% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{readDistMat}
\alias{readDistMat}
\title{read full distance matrix}
\usage{
readDistMat(fileName)
}
\arguments{
\item{fileName}{the input file that contains full n x n matrix (n rows and n columns)}
}
\value{
distMat a n x n matrix of pairwise distance between a set of n queries
}
\description{
read full distance matrix
}
\author{
Duc Luong
}

