% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{readFeatureVectors}
\alias{readFeatureVectors}
\title{read feature vector file in sparse representation}
\usage{
readFeatureVectors(fileName)
}
\arguments{
\item{fileName}{the input file with 3 columns (row, col, val)}
}
\value{
a n x p matrix where n is the number of queries and p is the number of features
}
\description{
read feature vector file in sparse representation
}
\author{
Duc Luong
}

