% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{readLabelFile}
\alias{readLabelFile}
\title{read the label file that contains the list of labels which outputs from WL algorithm}
\usage{
readLabelFile(filename)
}
\arguments{
\item{filename}{file that contains label file}
}
\value{
the list of labels
}
\description{
read the label file that contains the list of labels which outputs from WL algorithm
}
\author{
Duc Luong
}

