% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{readSparseDistMat}
\alias{readSparseDistMat}
\title{read sparse distance matrix (not yet implement)}
\usage{
readSparseDistMat(fileName)
}
\arguments{
\item{fileName}{the input file that contains matrix in sparse representation (csv file with header row, col, val)}
}
\value{
distMat a n x n matrix of pairwise distance between a set of n queries
}
\description{
read sparse distance matrix (not yet implement)
}
\author{
Duc Luong
}

