% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{silhouettePlot}
\alias{silhouettePlot}
\title{produce silhoette plot of the sql query similarity with their respective labels}
\usage{
silhouettePlot(distMat, label, fileName)
}
\arguments{
\item{distMat}{a n x n matrix of pairwise distance between a set of n queries}

\item{label}{a vector of size n that contains the labels of n queries}

\item{fileName}{the output file of the silhoutte plot}
}
\description{
produce silhoette plot of the sql query similarity with their respective labels
}
\author{
Duc Luong
}

