% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{sumWeights}
\alias{sumWeights}
\title{compute the sum of weight of all the connections between two clusters}
\usage{
sumWeights(distMat, label, i, j)
}
\arguments{
\item{distMat}{a n x n matrix of pairwise distance between a set of n queries}

\item{label}{a vector of size n that contains the labels of n queries}

\item{i, j}{cluster labels that we want to compute the sum of weights of all the edges between cluster i and cluster j}
}
\value{
the sum of weights of all the connections between two clusters
}
\description{
compute the sum of weight of all the connections between two clusters
}
\author{
Duc Luong
}

